remove hover; fixed vet renaming
