+++ modules/global/src/com/haulmont/sample/petclinic/entity/veterinarian/messages.properties
