+++ modules/global/src/com/haulmont/sample/petclinic/views.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <views xmlns="http://schemas.haulmont.com/cuba/view.xsd">
-    <view class="com.haulmont.sample.petclinic.entity.vet.Veterinarian"
+    <view class="com.haulmont.sample.petclinic.entity.veterinarian.Veterinarian"
           extends="_local"
-          name="vet-with-specialties">
+          name="veterinarian-with-specialties">
         <property name="specialties"
                   view="_minimal"/>
     </view>
