+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/veterinarian/specialty/messages.properties
