+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/veterinarian/veterinarian/messages.properties
