+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/veterinarian/veterinarian/veterinarian-browse.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <window xmlns="http://schemas.haulmont.com/cuba/screen/window.xsd"
         caption="msg://browseCaption"
-        focusComponent="vetsTable"
-        messagesPack="com.haulmont.sample.petclinic.web.vet.vet">
+        focusComponent="veterinariansTable"
+        messagesPack="com.haulmont.sample.petclinic.web.screens.veterinarian.veterinarian">
     <data readOnly="true">
-        <collection id="vetsDc"
-                    class="com.haulmont.sample.petclinic.entity.vet.Veterinarian"
-                    view="vet-with-specialties">
-            <loader id="vetsDl">
+        <collection id="veterinariansDc"
+                    class="com.haulmont.sample.petclinic.entity.veterinarian.Veterinarian"
+                    view="veterinarian-with-specialties">
+            <loader id="veterinariansDl">
                 <query>
                     <![CDATA[select e from petclinic_Veterinarian e]]>
                 </query>
@@ -16,15 +16,15 @@
     </data>
     <dialogMode height="600"
                 width="800"/>
-    <layout expand="vetsTable"
+    <layout expand="veterinariansTable"
             spacing="true">
         <filter id="filter"
-                applyTo="vetsTable"
-                dataLoader="vetsDl">
+                applyTo="veterinariansTable"
+                dataLoader="veterinariansDl">
             <properties include=".*"/>
         </filter>
-        <groupTable id="vetsTable"
-                    dataContainer="vetsDc"
+        <groupTable id="veterinariansTable"
+                    dataContainer="veterinariansDc"
                     width="100%">
             <actions>
                 <action id="create" type="create" />
@@ -42,9 +42,9 @@
             <rowsCount/>
             <buttonsPanel id="buttonsPanel"
                           alwaysVisible="true">
-                <button id="createBtn" action="vetsTable.create"/>
-                <button id="viewBtn" action="vetsTable.view"/>
-                <button id="removeBtn" action="vetsTable.remove"  stylename="danger"/>
+                <button id="createBtn" action="veterinariansTable.create"/>
+                <button id="viewBtn" action="veterinariansTable.view"/>
+                <button id="removeBtn" action="veterinariansTable.remove"  stylename="danger"/>
             </buttonsPanel>
         </groupTable>
         <hbox id="lookupActions" spacing="true" visible="false">
