+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/veterinarian/veterinarian/veterinarian-edit.xml
@@ -2,11 +2,11 @@
 <window xmlns="http://schemas.haulmont.com/cuba/screen/window.xsd"
         caption="msg://editorCaption"
         focusComponent="fieldGroup"
-        messagesPack="com.haulmont.sample.petclinic.web.vet.vet">
+        messagesPack="com.haulmont.sample.petclinic.web.screens.veterinarian.veterinarian">
     <data>
-        <instance id="vetDc"
-                  class="com.haulmont.sample.petclinic.entity.vet.Veterinarian"
-                  view="vet-with-specialties">
+        <instance id="veterinarianDc"
+                  class="com.haulmont.sample.petclinic.entity.veterinarian.Veterinarian"
+                  view="veterinarian-with-specialties">
             <loader/>
 
             <collection id="specialtiesDc" property="specialties"/>
@@ -18,7 +18,7 @@
                 width="800"/>
 
     <layout expand="editActions" spacing="true">
-        <form id="fieldGroup" dataContainer="vetDc">
+        <form id="fieldGroup" dataContainer="veterinarianDc">
             <column width="250px">
                 <textField property="firstName"/>
                 <textField property="lastName"/>
@@ -35,7 +35,7 @@
                     <action id="add" type="add">
                         <properties>
                             <property name="openMode" value="DIALOG"/>
-                            <property name="screenClass" value="com.haulmont.sample.petclinic.web.screens.vet.specialty.SpecialtyLookup"/>
+                            <property name="screenClass" value="com.haulmont.sample.petclinic.web.screens.veterinarian.specialty.SpecialtyLookup"/>
                         </properties>
                     </action>
                     <action id="remove" type="remove"/>
