--- modules/web/src/com/haulmont/sample/petclinic/web/vet/vet/VetEdit.java
@@ -1,14 +1,0 @@
-package com.haulmont.sample.petclinic.web.vet.vet;
-
-import com.haulmont.cuba.gui.Route;
-import com.haulmont.cuba.gui.screen.*;
-import com.haulmont.sample.petclinic.entity.vet.Veterinarian;
-
-@UiController("petclinic_Veterinarian.edit")
-@UiDescriptor("vet-edit.xml")
-@EditedEntityContainer("vetDc")
-@LoadDataBeforeShow
-@Route(value = "vets/edit", parentPrefix = "vets")
-public class VetEdit extends StandardEditor<Veterinarian> {
-
-}
