--- modules/web/themes/hover/com.haulmont.sample.petclinic/app-component.scss
@@ -1,5 +1,0 @@
-@import "hover-ext";
-
-@mixin com_haulmont_sample_petclinic {
-  @include com_haulmont_sample_petclinic-hover-ext;
-}
