--- modules/web/themes/hover/com.haulmont.sample.petclinic/hover-ext-defaults.scss
@@ -1,24 +1,0 @@
-/* Override global theme constants or define new ones here */
-
-
-$petclinic-light-color: #badff3;
-
-$petclinic-main-color: #00a1d9;
-
-$petclinic-main-bg: #f1faff;
-
-$petclinic-login-btn: #0093cf;
-
-
-
-$v-focus-color: $petclinic-main-color;
-
-$v-selection-color: $petclinic-main-color;
-
-$v-border: 1px solid $petclinic-light-color;
-
-$v-friendly-color: $petclinic-main-color;
-
-$v-font-color: $petclinic-main-color;
-
-$v-button-font-color: #000;
