--- modules/web/themes/hover/com.haulmont.sample.petclinic/hover-ext.scss
@@ -1,4 +1,0 @@
-/* Define your theme modifications inside next mixin */
-@mixin com_haulmont_sample_petclinic-hover-ext {
-
-}
