--- modules/web/themes/hover/styles.scss
@@ -1,215 +1,0 @@
-@import "hover-defaults";
-@import "com.haulmont.sample.petclinic/hover-ext-defaults";
-@import "app-components";
-@import "com.haulmont.sample.petclinic/hover-ext";
-
-@mixin calendar-event($event-name, $text-color, $background-color) {
-
-    .v-calendar-event.#{$event-name} {
-      background-color: $background-color;
-      color: $text-color;
-    }
-    // Weekly view event text color, Monthly view event 'point' color
-    .v-calendar-week-wrapper .v-calendar-event-#{$event-name}, .v-calendar-event-#{$event-name}:before {
-      color: $text-color;
-    }
-    // Weekly view event background color
-    .v-calendar-week-wrapper .v-calendar-event-#{$event-name} .v-calendar-event-content {
-      background-color: $background-color;
-    }
-}
-
-.hover {
-  // include auto-generated app components SCSS
-  @include app_components;
-  
-  @include com_haulmont_sample_petclinic-hover-ext;
-
-  @include calendar-event(event-red, #e00000, #f4c8c8);
-  @include calendar-event(event-green, #3c853a, #c8f4ca);
-  @include calendar-event(event-yellow, #7f7e76, #f4efc8);
-  @include calendar-event(event-purple, #ffffff, #b411dd);
-  @include calendar-event(event-blue, #00a2d9, #bae0f3);
-
-  .c-groupbox-caption {
-    color: $petclinic-main-color;
-  }
-
-
-  .c-login-submit-button {
-    background-color: $petclinic-login-btn;
-  }
-
-  .v-slot-c-sidemenu-panel {
-    border:none;
-    background-color: $petclinic-light-color;
-
-  }
-
-  .v-textfield-focus, .v-textfield:focus {
-    border-color: $petclinic-main-color;
-  }
-
-
-  //////////////////////////////////////////////////////////////////////////////////////////////////
-  // Side menu
-  //////////////////////////////////////////////////////////////////////////////////////////////////
-  .c-sidemenu-panel {
-
-    background: #ffffff none;
-    border-right: none;
-    color: $petclinic-main-color;
-
-
-    .c-sidemenu-item-action.c-sidemenu-item-focused {
-      border-right: 5px solid $petclinic-main-color;
-    }
-    .c-sidemenu-item-header, .c-sidemenu-item-header :hover {
-      color: $petclinic-main-color;
-      margin: 0;
-    }
-
-    .c-sidemenu-item-header-open {
-      background-color: $petclinic-main-color;
-      color:#ffffff;
-    }
-    .c-sidemenu-submenu-open {
-      background-color: $petclinic-light-color;
-      color:$petclinic-main-color;
-    }
-
-    .c-sidemenu-item :focus, .c-sidemenu-item :hover, .c-sidemenu-item-focused {
-      color: #ffffff;
-    }
-
-    .c-sidemenu-item-header.c-sidemenu-item-focused, .c-sidemenu-item-selected {
-      background-color: $petclinic-main-color;
-      color: #ffffff;
-    }
-
-    .c-sidemenu {
-      color: $petclinic-main-color;
-    }
-
-    .c-sidemenu-title {
-      color: #ffffff;
-      background-color: $petclinic-main-color;
-
-    }
-
-    .c-ftsfield-button, .c-newwindow-button, .c-logout-button {
-      color: $petclinic-main-color;
-    }
-
-    .c-ftsfield-button:hover, .c-newwindow-button:hover, .c-logout-button:hover {
-      color: $petclinic-main-color;
-      background: transparent;
-    }
-
-  }
-
-  .v-select-optiongroup-button-switcher {
-
-      @include valo-optiongroup-horizontal;
-
-      .v-radiobutton.v-select-option {
-        margin: 0;
-        padding-left: 0;
-      }
-
-      .v-radiobutton.v-select-option > input {
-
-        &:checked ~ label {
-          @include valo-button-style($background-color: $v-focus-color, $unit-size: null);
-        }
-
-        & ~ label:after,
-        &:checked ~ label:after {
-          width: 100%;
-          height: 100%;
-          top: 0;
-          left: 0;
-          border-radius: 0;
-          border: 1px solid;
-          background-color: transparent;
-        }
-
-        &:checked:hover ~ label:after {
-          @include valo-button-hover-style($background-color: $v-focus-color);
-        }
-
-        &:checked:focus ~ label:after {
-          $checked-text-color: valo-font-color($v-focus-color, 0.9);
-          $focus-style: inset 0px 0px 3px 1px rgba($checked-text-color, 0.5);
-          @include box-shadow($focus-style);
-        }
-
-        &:checked:active ~ label:after {
-          @include valo-button-active-style($background-color: $v-focus-color);
-        }
-
-        &:hover ~ label:after {
-          @include valo-button-hover-style($background-color: $v-background-color);
-        }
-
-        &:focus ~ label:after {
-          @include valo-button-focus-style();
-        }
-
-        &:active ~ label:after {
-          @include valo-button-active-style($background-color: $v-background-color);
-        }
-
-        $padding-height: 6px;
-        $padding-width: round($v-unit-size/2.4) + round($v-border-radius/3);
-        $border-width: first-number($v-border);
-        $height: $v-unit-size - 2 * $padding-height - 2 * $border-width;
-
-        & ~ label {
-          height: $height;
-          padding: $padding-height $padding-width;
-          border-radius: 0;
-          @include valo-border-with-gradient($border: $v-border, $color: $v-background-color, $gradient: $v-gradient);
-          @include valo-gradient($v-background-color, $v-gradient);
-          @include box-shadow(valo-bevel-and-shadow($bevel: $v-bevel, $shadow: $v-shadow, $background-color: $v-background-color, $gradient: $v-gradient));
-        }
-
-        & ~ label:before {
-          display: none;
-        }
-      }
-
-      .v-radiobutton.v-select-option:first-child,
-      .v-radiobutton.v-select-option:last-child {
-        margin: 0;
-      }
-
-      .v-radiobutton.v-select-option:first-child > input {
-        & ~ label {
-          border-top-left-radius: $v-border-radius;
-          border-bottom-left-radius: $v-border-radius;
-
-          &:after {
-            border-top-left-radius: $v-border-radius;
-            border-bottom-left-radius: $v-border-radius;
-          }
-        }
-      }
-
-      .v-radiobutton.v-select-option:last-child > input {
-        & ~ label {
-          border-top-right-radius: $v-border-radius;
-          border-bottom-right-radius: $v-border-radius;
-
-          &:after {
-            border-top-right-radius: $v-border-radius;
-            border-bottom-right-radius: $v-border-radius;
-          }
-        }
-      }
-
-      .v-checkbox > input ~ label:after {
-        content: "";
-      }
-    }
-}
